DRY out trailer-based detection into single package
